rag_application/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic[email]==2.9.2
email-validator==2.2.0
pydantic-settings==2.5.2
python-multipart==0.0.12

sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

chromadb==0.5.15
google-generativeai==0.8.3
pymupdf==1.24.10

httpx==0.27.2

pytest==8.3.3
pytest-mock==3.14.0
pytest-cov==5.0.0
pytest-asyncio==0.24.0
